Add files via upload
Small change in wording when waiting for a connection; doesn't say that it restarts the server each time a user disconnects, but rather that it is waiting for a connection. This has been bugging me so I fixed it. It may also contain minor bug fixes, but I can't remember if I did.